Better tests & documentatioin & working stash reference
