+++ go/libraries/doltcore/doltdb/stash.go
@@ -26,14 +26,15 @@ import (
 )
 
 type Stash struct {
-	Name        string
-	BranchName  string
-	Description string
-	HeadCommit  *Commit
+	Name           string
+	BranchName     string
+	Description    string
+	HeadCommit     *Commit
+	StashReference string
 }
 
 // getStashList returns array of Stash objects containing all stash entries in the stash list map.
-func getStashList(ctx context.Context, ds datas.Dataset, vrw types.ValueReadWriter, ns tree.NodeStore) ([]*Stash, error) {
+func getStashList(ctx context.Context, ds datas.Dataset, vrw types.ValueReadWriter, ns tree.NodeStore, reference string) ([]*Stash, error) {
 	v, ok := ds.MaybeHead()
 	if !ok {
 		return nil, errors.New("stashes not found")
@@ -75,6 +76,7 @@ func getStashList(ctx context.Context, ds datas.Dataset, vrw types.ValueReadWrit
 		s.HeadCommit = headCommit
 		s.BranchName = meta.BranchName
 		s.Description = meta.Description
+		s.StashReference = reference
 
 		sl[i] = &s
 	}
